Refactor UsingURLPatterns to use override_settings(ROOT_URLCONF=...) style
